requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
google-genai==1.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
